feature/KAN-60: End-to-End Test (#60)
* KAN-60: End-to-End Test

created a new test file for end-to-end testing
made changes to the rides schema to include payment_charged as a float instead of int aligning with the fleet manager
inserted failsafe for the end time in the fleet manager to ensure it is always after the start time fixing inconsistencies in the tests

* Updated the status code for user registration endpoint to 201-Created
Updated the test for user registration

---------

+++ src/api/schemas/rides.py
@@ -23,7 +23,7 @@ class EndRideRequest(StrictBaseModel):
 class EndRideResponse(StrictBaseModel):
     ride_id: int
     end_station_id: int
-    payment_charged: int  # rename from payment_charged_ils
+    payment_charged: float  # rename from payment_charged_ils
 
 
 class ActiveUsersResponse(StrictBaseModel):
